requirements.txt
asgiref==3.5.1
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
decorator==5.1.1
Django==4.0.4
django-bootstrap-datepicker-plus==4.0.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-filter==21.1
djangorestframework==3.13.1
executing==0.8.3
idna==3.3
ipdb==0.13.9
ipython==8.3.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.29
psycopg2==2.9.3
pure-eval==0.2.2
Pygments==2.12.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
traitlets==5.2.0
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5
workadays==2021.12.18
